Move functions in main.py to src/runner.py
